Adds the IBaseApplicationService
This interface declares the functionalities common to all the objects
in the BusinessLayer. Have to be inherits from the others interfaces
that will contain the functionalities of a given Entity.

NOTE:
This implementation is the standard for the projects that I implement with the
Multi-Layer architecture. If any improvement is found, you can suggest it.