Implements the class that is gonna be used by the unitTests to resolve the Context
For each set of unitTests, we have to use a context, this is context can be sqlLiteInMemory
which is the best option because is the faster one, but there are some instructions os SqlServer
that are not valid in sqlLite, that is why sometimes we'll have to use a DB in SQlServer.

This class that is committed, creates the context, apply the migrations and returns the context. After
the context is not needed no more, the db is deleted