Fixes error in the creation of the Db
It seems that the method context.Database.EnsureCreate() has some issues
that I can't identify now. Using the method Database.Migrate() there is not
problem in the creation of the DB because this method runs the migrations.

NOTE_1:
I changed the connectionString. In case that you don't have the default sqlServer
we have agree about the name of the server.

NOTE_2:
This state fixes some unitTest method. Have to change the code of all the unitTests
to not assign the Id to the new objects.