Fixes bug with workaround
Now after initialize the DB, we are asking if the Table ThoughtHoles contains
the column CreatedAt, if it does, then do nothing, if does not, then adds the column.

This can't not be used for sqlLite. It seems there is a bug with the sqlLite of EF Core, because
after the first creation of the DB it says that we are adding a duplicate column. So, we are gonna
use SqlServer by default till this bug will get fix