Implement functionalities to apply all the seeds using reflection
Because the seeds have to be implemented in the same assembly where the
user declare the custom appDbContext, the method get all the types in this
assembly that implement the interface `ISeed` and because all this members have
the same method `AddOrUpdate`, it's possible to call it with reflection

NOTE:
Was trying to implement a way to declare the order in which the seeds have
to be called. The first approch is having a property in each seed class, but with
reflection is not possible to get its value because the class is not instantiated. The
property has to be static to get the value, but the interfaces can't declare static
members, so this is not possible.

Gonna try this using attributes later.