Remove the  context param from the extension method
I had to pass the context, because there was an error resolving its
dependency. Now, I changed the config of the service in the
ConfigureServices method, the context and the DbOptions are transient,
which means that they are created every time they are requested. The
problem with the scoped is that it can't be used for this, because there is
not request when we run the tests.

NOTE: Should create another method that recieve the DBContext in case
that the users want the DbContext as scoped.